requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.0
pydantic-settings==2.4.0

# Document parsing
docling==2.5.0

# Vector DB
qdrant-client==1.11.3

# LLM / HTTP
openai==1.51.0
httpx==0.27.2

# Chunking
langchain-text-splitters==0.3.0

# Evaluation
ragas==0.1.21
datasets==2.21.0

# Utils
python-multipart==0.0.9
pyyaml==6.0.2
structlog==24.4.0
python-dotenv==1.0.1
tqdm==4.66.5
